refactor: 콤보 랭킹 조회 서비스 분리
