+++ src/main/kotlin/com/example/kbocombo/combo/application/ComboService.kt
@@ -2,9 +2,6 @@ package com.example.kbocombo.combo.application
 
 import com.example.kbocombo.combo.application.request.ComboCreateRequest
 import com.example.kbocombo.combo.domain.Combo
-import com.example.kbocombo.combo.domain.ComboFailedEvent
-import com.example.kbocombo.combo.domain.ComboPassedEvent
-import com.example.kbocombo.combo.domain.ComboSucceedEvent
 import com.example.kbocombo.combo.domain.vo.ComboStatus
 import com.example.kbocombo.combo.infra.ComboRepository
 import com.example.kbocombo.combo.infra.getById
@@ -66,7 +63,6 @@ class ComboService(
         combos.forEach {
             it.success()
             comboRepository.save(it)
-            publisher.publishEvent(ComboSucceedEvent(memberId = it.memberId, comboId = it.id))
         }
     }
 
@@ -85,7 +81,6 @@ class ComboService(
         combos.forEach {
             it.fail()
             comboRepository.save(it)
-            publisher.publishEvent(ComboFailedEvent(memberId = it.memberId, comboId = it.id))
         }
     }
 
@@ -102,7 +97,6 @@ class ComboService(
         combos.forEach {
             it.pass()
             comboRepository.save(it)
-            publisher.publishEvent(ComboPassedEvent(memberId = it.memberId, comboId = it.id))
         }
     }
